generate type index
DWARF 5 and the pre-standard GNU extension expect different sections to
exist and has different formats for the `.debug_{cu,tu}_index`
sections - support for these differences is clarified and improved. In
addition, split type units (in DWARF 5) and `.debug_types` (in DWARF 4)
are now supported, resulting in generation of a `.debug_tu_index`
section.

Signed-off-by: David Wood <david.wood@huawei.com>