refactor into methods on `OutputPackage`
Refactor `append_*` functions into methods on `OutputPackage` and add
additional fields for the in-progress structures being accumulated into
during the appending (e.g. string table or index entries). Enables
significant simplification of the signatures of these functions as all
of the output variables no longer need to be passed around.

Signed-off-by: David Wood <david.wood@huawei.com>