adjust contributions of units from input dwp files
Contributions of individual units in input DWARF packages need to be
preserved. For a input DWARF object, the contribution of a unit to a
non-unit section (like `.debug_str_offsets`) is the entirety of the
other section. However, with an input DWARF package, its units'
contributions are parts of the non-unit sections in the DWARF package.

Contributions from units in dwp files are now adjusted based on their
entry in the input files' index (with the offset including the sizes
of the previous units, but relative to its position in the new output
package).

Signed-off-by: David Wood <david.wood@huawei.com>